Reorganized project directory structure to have a proper package name (instead of src). Added setup.py for easy installation.
