+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(name="SigasiProjectCreator",
+      description="Generate a Sigasi Project from your own project specifications",
+      url='https://github.com/sigasi/SigasiProjectCreator',
+      packages=['SigasiProjectCreator'],
+      package_dir={'': 'src'},
+      )
