+++ src/SigasiProjectCreator/ConverterHelper.py
