+++ src/SigasiProjectCreator/createSigasiProjectFromListOfFiles.py
